Add LGet/LSet/LSetWithFlags/LList/LRemove
This commit adds "L" variants for all public functions.
These do not follow a symlink at the end of the path.

All platforms provide a system call interface for that,
even though it looks different, and this patch adds the
appropriate calls:

* lgetxattr/lsetxattr/llistxattr on Linux
* SYS_EXTATTR_GET_LINK/SYS_EXTATTR_SET_LINK/SYS_EXTATTR_LIST_LINK on FreeBSD
* XATTR_NOFOLLOW on Darwin

For Linux, we use the syscall wrappers that the x/sys/unix package
provides. for FreeBSD and Darwin, new syscall.Syscall6 wrappers
are added as needed.

Tests are added for all new functions. I have tested on FreeBSD
and Linux.