support "F" variants of operations (on file descriptors)
In unix, there is a possibility to create files without a path, for
example by using O_TMPFILE flag in open(2).

This change adds a set of "F" variants to all functions that accepts
a file descriptor instead of a path.

The change is only implemented for Linux, all other platforms use Name()
call on a file to get its path.