Fix FIFO handling on Solaris (#70)
* Fix FIFO handling on Solaris

On Solaris, files need to be opened to obtain or change their
attributes. But opening a FIFO file causes the calling thread to block
until the write end is opened. Files are now opened using O_NONBLOCK to
prevent this from happening.

Should fix restic/restic#4003.

Also fix comment on stringsFromByteSlice.

* (Solaris) Open O_NONBLOCK with unix.Open, not os.Open

Passing this flag to os.Open works, but this relies on undocumented
implementation details of package os.

* (Solaris) Fix removexattr/hanging test